cleanup file names and improve tests
